Move `_load_library` to `torchvision.extension`
